requirements.txt
tkinter==8.6.12
ttkbootstrap==0.5.1
PyPDF2==3.0.1
pdf2image==1.16.3
Pillow==9.5.0